test(filewatcher): merge scripted tests into atelier-test
`runFileWatcherScripted` lives in atelier, so its tests belong there.
Removes `Unit.Tricorder.FileWatcherSpec` and folds the scripted
interpreter cases into `Unit.Atelier.Effects.FileWatcherSpec`.
