+++ atelier/test/Unit/Atelier/Effects/FileWatcherSpec.hs
@@ -1,6 +1,10 @@
 module Unit.Atelier.Effects.FileWatcherSpec (spec_FileWatcher) where
 
+import Control.Concurrent (forkIO, killThread, newQSem, signalQSem, waitQSem)
+import Data.IORef (modifyIORef, newIORef, readIORef)
 import Data.List (isSuffixOf)
+import Effectful (IOE, runEff)
+import Effectful.Concurrent (Concurrent, runConcurrent)
 import Hedgehog (Gen, PropertyT, forAll, (===))
 import Test.Hspec (Spec, describe, it, shouldBe)
 import Test.Hspec.Hedgehog (hedgehog)
@@ -8,7 +12,7 @@ import Test.Hspec.Hedgehog (hedgehog)
 import Hedgehog.Gen qualified as Gen
 import Hedgehog.Range qualified as Range
 
-import Atelier.Effects.FileWatcher (deduplicateDirs, dir, dirWhere, matchesAny)
+import Atelier.Effects.FileWatcher (FileWatcher, Watch, deduplicateDirs, dir, dirWhere, matchesAny, runFileWatcherScripted, watchFilePaths)
 
 
 spec_FileWatcher :: Spec
@@ -47,6 +51,59 @@ spec_FileWatcher = do
                 `shouldBe` True
             matchesAny [dirWhere "/proj/src" (\f -> ".hs" `isSuffixOf` f)] "/proj/src/Foo.js"
                 `shouldBe` False
+
+    describe "runFileWatcherScripted" testScripted
+
+
+--------------------------------------------------------------------------------
+-- Scripted interpreter tests
+--------------------------------------------------------------------------------
+
+testScripted :: Spec
+testScripted = do
+    describe "watchFilePaths" do
+        it "calls the callback with the scripted path" do
+            paths <- collectPaths ["/src/Foo.hs"]
+            paths `shouldBe` ["/src/Foo.hs"]
+
+        it "calls the callback with each path in order" do
+            paths <- collectPaths ["/src/Foo.hs", "/src/Bar.hs"]
+            paths `shouldBe` ["/src/Foo.hs", "/src/Bar.hs"]
+
+        it "ignores the watch specification" do
+            paths <- collectPathsWith [dir "/any"] ["/src/Foo.hs"]
+            paths `shouldBe` ["/src/Foo.hs"]
+
+        it "passes the full path to the callback unchanged" do
+            let path = "/home/user/project/src/Some/Deep/Module.hs"
+            paths <- collectPaths [path]
+            paths `shouldBe` [path]
+
+
+--------------------------------------------------------------------------------
+-- Helpers
+--------------------------------------------------------------------------------
+
+-- | Run the scripted interpreter, collecting all callback-delivered paths.
+-- Uses a semaphore to wait for exactly N events before cancelling the watcher.
+collectPaths :: [FilePath] -> IO [FilePath]
+collectPaths = collectPathsWith []
+
+
+collectPathsWith :: [Watch] -> [FilePath] -> IO [FilePath]
+collectPathsWith watches scripted = do
+    ref <- newIORef []
+    sem <- newQSem 0
+    tid <- forkIO $ void $ runScripted scripted $ watchFilePaths watches \p -> liftIO do
+        modifyIORef ref (<> [p])
+        signalQSem sem
+    replicateM_ (length scripted) (waitQSem sem)
+    killThread tid
+    readIORef ref
+
+
+runScripted :: [FilePath] -> Eff '[FileWatcher, Concurrent, IOE] a -> IO a
+runScripted paths = runEff . runConcurrent . runFileWatcherScripted paths
 
 
 --------------------------------------------------------------------------------
